IDE-1136
Add quick fix for the project of which primary runtime has not been set.
If there is no available runtime, open the new Liferay Server wizard;
If there is only one, set it as the primary runtime directly;
If there are multiple runtims, then open the Targeted Runtimes property
page of the project.